--- a/sensor/kubernetes/listener/resources/convert_mutating_test.go
+++ b/sensor/kubernetes/listener/resources/convert_mutating_test.go
@@ -312,7 +312,7 @@ func TestConvertDifferentContainerNumbers(t *testing.T) {
 	for _, c := range cases {
 		c := c
 		t.Run(c.name, func(t *testing.T) {
-			actual := newDeploymentEventFromResource(c.inputObj, &c.action, c.deploymentType, testClusterID, c.podLister, mockNamespaceStore, hierarchyFromPodLister(c.podLister), "", c.systemNamespaces, c.registryStore).GetDeployment()
+			actual := newDeploymentEventFromResource(c.inputObj, &c.action, c.deploymentType, testClusterID, c.podLister, mockNamespaceStore, hierarchyFromPodLister(c.podLister), "", c.systemNamespaces).GetDeployment()
 			if actual != nil {
 				actual.StateTimestamp = 0
 			}