## Description

Rather than deduping entire deployments at the detector level, this proposal is to generated hashes out of the `store.Dependencies` object that is used to enrich the deployment spec into a `storage.Deployment`. If neither the deployment spec nor the 
dependencies changed, there is no reason to process a deployment further. 

**Overview**

On a deploy-time path (i.e. Kubernetes events) the only reasons to process a deployment on an `UPDATE` (send a `SensorEvent_Resource` message to central and check violations) is when either of these two happen:
1. The deployment spec changed
2. Some adjacent resource changed (e.g. RBAC or Route)

Whenever a deployment event is received, we store the spec in deployment store as `built=false`. Meaning that the present deployment is yet to be resolved against its dependencies. This deployment is then sent to the `resolver`, which will gather the dependencies (from RBAC, Pod and Services stores) and build the `storage.Deployment` object that will be sent to Central and to the detector.

Whenever an adjacent resource changes, a deployment reference is sent to the `resolver` as well. Signalling that the deployment has to be processed again since a dependency changed.

The way we make sure we don't over process deployments is by deduping the `storage.Deployment` object at the detector, after building it against its dependencies. However, this still incur some two costly operations: **hashing** and **cloning** the deployment object.

**The proposal**

The key change is how sensor identifies if a deployment has to be processed. At the `resolver` component, we have two key pieces of information that could be used to avoid processing the deployment: the `built` property, which determines if there were changes in the deployment spec since last processed. And the `dependencies` parameter, which are the properties that deployments will be enriched on. 

- Hash `dependencies` struct
- Introduce a snapshot map, which holds the latest `storage.Deployment` built and its dependency hash value
- Whenever a deployment reference is received, if `built=true` and the dependency hash is the same, there is no need to process the deployment.
- If the detection was forced (e.g. image cache updated or reprocess requested) grab the latest built deployment from the snapshot map and send to the detector.

**Hypothesis**

This change would make the deduper in the detector unnecessary. Since there are no reasons why a deployment should be processed against deploy-time policies if neither the spec nor the dependencies changed. Deployments will only be passed if there is a relevant change (which would then result in a dedupe check miss anyway) or if they are forced. 

**Example scenarios:**

|Scenario|Old behavior|Proposed behavior|
|---|---|---|
|Secret changed, but no deployment is affected|All deployments are processed by resolver; deployments are hashed and cloned; detector dedupes them and skip processing| All deployment processed by resolver; `store.Dependencies` are identical from previous built deployment; no deployments are processed|
|Service spec changed, but deployment is not affected|Matching deployment processed by resolver; deployments are hashed and cloned; detector dedupes them and skips processing | Matching deployment processed by resolver; `store.Dependencies` are identical from previous built deployment; no deployments are processed|
|Central requested all deployments to be reprocessed |All deployment processed by resolver; deduper reset|All deployment processed by resolver; deduper reset|
|Deployment spec changed|`isBuilt` is set to false; deployment fully processed|`isBuilt` is set to false; deployment fully processed|

**Results**

Reducing allocations by 98% on a path were multiple resource updates would trigger deployment that we don't care about.


## Checklist

- [x] Implement improvement (present PR)
- [x] Benchmark results
- [ ] ~~Validate that detector's deduper is no longer needed~~
- [ ] Compare pprof results from real workloads
- [ ] ~~Remove deduper from detector~~
- [x] Investigated and inspected CI test results
- [x] Unit test and regression tests added
- [ ] Evaluated and added CHANGELOG entry if required
- [ ] ~~Determined and documented upgrade steps~~
- [ ] ~~Documented user facing changes (create PR based on [openshift/openshift-docs](https://github.com/openshift/openshift-docs) and merge into [rhacs-docs](https://github.com/openshift/openshift-docs/tree/rhacs-docs))~~

If any of these don't apply, please comment below.

## Testing Performed

### Here I tell how I validated my change

Unit test cases were added to make sure that `BuildDeploymentWithDependencies` will return generate the new deployment object when necessary. 

**Benchmarks**

Two benchmark cases were ran, one where the deployment has to be reprocessed; and one where no reprocessing would be needed. In both test cases, 100 deployment updates were triggered.

|Deployment reprocessing required|Allocations (current)|Allocations (new)| Allocations (new + removing hashes)|
|---|---|---|---|
|No|14MB|520K (-96%)|511K(-96%)|
|Yes|14MB|15MB (+3%)|916K(-93%)|

Notice that the current implementation makes the case where updates are always relevant worse by 3%. This is because we hash the dependencies store **and** the entire deployment object. 

Benchmark results:

```pkg: github.com/stackrox/rox/sensor/kubernetes/listener/resources
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
                             │ ./benchtest_build_deployment_old.txt │  benchtest_build_deployment_new.txt  │ ./benchtest_build_deployment_new_no_hash.txt │
                             │                sec/op                │    sec/op      vs base               │        sec/op          vs base               │
BuildDeployments_NoChange-12                           863.86m ± 2%    38.87m ± 12%  -95.50% (p=0.002 n=6)             34.27m ± 1%  -96.03% (p=0.002 n=6)
BuildDeployments_Change-12                             861.58m ± 2%   895.38m ±  2%   +3.92% (p=0.002 n=6)             66.50m ± 1%  -92.28% (p=0.002 n=6)
geomean                                                 862.7m         186.6m        -78.38%                           47.74m       -94.47%

                             │ ./benchtest_build_deployment_old.txt │  benchtest_build_deployment_new.txt  │ ./benchtest_build_deployment_new_no_hash.txt │
                             │                 B/op                 │     B/op       vs base               │         B/op           vs base               │
BuildDeployments_NoChange-12                         237.349Mi ± 0%    7.690Mi ± 0%  -96.76% (p=0.002 n=6)            7.565Mi ± 0%  -96.81% (p=0.002 n=6)
BuildDeployments_Change-12                            237.38Mi ± 0%   244.85Mi ± 0%   +3.15% (p=0.002 n=6)            41.62Mi ± 0%  -82.47% (p=0.002 n=6)
geomean                                                237.4Mi         43.39Mi       -81.72%                          17.74Mi       -92.52%

                             │ ./benchtest_build_deployment_old.txt │  benchtest_build_deployment_new.txt  │ ./benchtest_build_deployment_new_no_hash.txt │
                             │              allocs/op               │   allocs/op    vs base               │       allocs/op        vs base               │
BuildDeployments_NoChange-12                          14777.2k ± 0%     520.0k ± 0%  -96.48% (p=0.002 n=6)             511.1k ± 0%  -96.54% (p=0.002 n=6)
BuildDeployments_Change-12                            14779.0k ± 0%   15286.6k ± 0%   +3.43% (p=0.002 n=6)             916.7k ± 0%  -93.80% (p=0.002 n=6)
geomean                                                 14.78M          2.819M       -80.92%                           684.5k       -95.37%
```

**Detector deduper**

It's currently not possible to remove the deduper because there are pod updates that will cause deployments to be processed (with `isBuilt=false` since spec changed) but the resulting hash will be the same. Fixing this will require a more elaborate change in the deployment handler.

### Reminder for reviewers

In addition to reviewing code here, reviewers **must** also review testing and request further testing in case the
performed one does not seem sufficient. As a reviewer, you must not approve the change until you understand the
performed testing and you are satisfied with it.